Rename connect package to rp (RChain Protocol)
